+++ reikna/cluda/cuda.py
@@ -79,7 +79,9 @@ class Array(gpuarray.GPUArray):
         Unlike PyOpenCL, PyCUDA's copy() does not use _new_like_me(),
         so we're overriding it.
         """
-        return self._new_like_me()
+        new_arr = self._new_like_me()
+        gpuarray._memcpy_discontig(new_arr, self, async=True, stream=self.thread._queue)
+        return new_arr
 
     def _new_like_me(self, dtype=None):
         """
@@ -262,5 +264,5 @@ class Kernel(api_base.Kernel):
         self._grid = tuple(grid) + (1,) * (3 - len(grid))
 
     def prepared_call(self, *args):
-        self._kernel(*args, grid=self._grid, block=self._local_size,
+        return self._kernel(*args, grid=self._grid, block=self._local_size,
             stream=self._thr._queue, shared=self._local_mem)